requirements.txt
langchain==0.2.16
langchain-core==0.2.41
langchain-openai==0.1.23
langchain-pinecone==0.1.3
pinecone==5.4.0
openai==1.40.6
python-dotenv==1.0.1
tiktoken==0.7.0